ContentSearchService/requirements.txt
fastapi
uvicorn
pymongo
dnspython
python-dotenv
requests
py-eureka-client
pydantic
httpx
